refactor(client): step 4 -- peel configs + queue families into mixins (#520)
Pure movement. Extracts two mid-sized families into per-family mixins over
_CoreClient:

  client/configs.py  _ConfigsMixin  (17 methods: components, configs, rows,
                     metadata, incl. delete_config)
  client/queue.py    _QueueMixin    (7 methods: job list/create/kill/events/poll)

Every method body, signature and docstring moved verbatim; each module gets
exactly the imports its methods use. No cross-family self-calls in either.

_client.py: 2258 -> 1508 LOC. make check green: 4648 passed, 8 skipped. ty clean.
