+++ src/keboola_agent_cli/client/_client.py
@@ -7,9 +7,7 @@ mapping are encapsulated here.
 Inherits shared retry/error logic from BaseHttpClient.
 """
 
-import json
 import logging
-import time
 from pathlib import Path
 from typing import Any
 from urllib.parse import quote
@@ -18,17 +16,11 @@ import httpx
 
 from ..constants import (
     CLOUD_UPLOAD_ERROR_BODY_LIMIT,
-    DEFAULT_GROUPED_JOBS_LIMIT,
-    DEFAULT_JOB_LIMIT,
-    DEFAULT_JOBS_PER_CONFIG,
-    DEFAULT_POLL_STRATEGY,
     EXPORT_JOB_MAX_WAIT,
     FILE_DOWNLOAD_CHUNK_SIZE,
     FILE_DOWNLOAD_TIMEOUT,
     FILE_UPLOAD_TIMEOUT,
     IMPORT_JOB_MAX_WAIT,
-    STORAGE_JOB_MAX_WAIT,
-    VALID_POLL_STRATEGIES,
 )
 from ..errors import ErrorCode, KeboolaApiError
 from ._core import _CoreClient
@@ -37,13 +29,14 @@ from ._transfer import (
     _build_abs_upload_url,
     _CloudDownloader,
     _extract_cloud_error_code,
-    _iter_poll_intervals,
     _IterBytesReader,
     _s3_signed_headers,
 )
 from .branches import _BranchesMixin
+from .configs import _ConfigsMixin
 from .misc import _MiscMixin
 from .query import _QueryMixin
+from .queue import _QueueMixin
 from .stream import _StreamMixin
 from .tokens import _TokensMixin
 from .workspaces import _WorkspacesMixin
@@ -52,6 +45,8 @@ logger = logging.getLogger(__name__)
 
 
 class KeboolaClient(
+    _QueueMixin,
+    _ConfigsMixin,
     _MiscMixin,
     _StreamMixin,
     _BranchesMixin,
@@ -68,481 +63,6 @@ class KeboolaClient(
 
     Inherits _do_request() and _raise_api_error() from BaseHttpClient.
     """
-
-    def list_components(
-        self,
-        component_type: str | None = None,
-        branch_id: int | None = None,
-    ) -> list[dict[str, Any]]:
-        """List components with their configurations.
-
-        Args:
-            component_type: Optional filter (extractor, writer, transformation, application).
-            branch_id: If set, list components from a specific dev branch.
-
-        Returns:
-            List of component dicts from the API.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        params: dict[str, str] = {"include": "configuration"}
-        if component_type:
-            params["componentType"] = component_type
-
-        response = self._request("GET", f"{prefix}/components", params=params)
-        return response.json()
-
-    def list_components_with_configs(
-        self,
-        branch_id: int | None = None,
-        component_type: str | None = None,
-        include_state: bool = False,
-    ) -> list[dict[str, Any]]:
-        """List all components with full configuration bodies and rows.
-
-        Makes a single API call to fetch everything needed for sync pull and
-        for deep search (row-level configuration). Uses the
-        include=configuration,rows parameter to get full config bodies and
-        config rows in one request. When ``include_state`` is True, the
-        response also embeds each configuration's runtime ``state`` dict
-        (same data as ``get_config_state``) so bulk-state retrieval stays a
-        single request instead of N+1. Also used by the bulk-detail caller
-        in ``ConfigService`` when ``--with-state`` is set on
-        ``config detail`` without a specific ``--config-id``.
-
-        Args:
-            branch_id: If set, target a specific dev branch.
-            component_type: Optional filter (extractor, writer, transformation,
-                application). Passed to the API as ``componentType``.
-            include_state: When True, adds ``state`` to the ``include``
-                resource list so each returned configuration carries its
-                runtime state dict.
-
-        Returns:
-            List of component dicts, each containing a 'configurations' list
-            with full config bodies and nested 'rows'.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        include_parts = ["configuration", "rows"]
-        if include_state:
-            include_parts.append("state")
-        params: dict[str, str] = {"include": ",".join(include_parts)}
-        if component_type:
-            params["componentType"] = component_type
-        resp = self._request(
-            "GET",
-            f"{prefix}/components",
-            params=params,
-        )
-        return resp.json()
-
-    def list_component_configs(
-        self,
-        component_id: str,
-        branch_id: int | None = None,
-    ) -> list[dict[str, Any]]:
-        """List all configurations for a specific component.
-
-        Args:
-            component_id: Component identifier (e.g. 'keboola.sandboxes').
-            branch_id: If set, target a specific dev branch.
-
-        Returns:
-            List of configuration dicts (id, name, description, etc.).
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        resp = self._request(
-            "GET",
-            f"{prefix}/components/{quote(component_id, safe='')}/configs",
-        )
-        return resp.json()
-
-    def list_config_rows(
-        self,
-        component_id: str,
-        config_id: str,
-        branch_id: int | None = None,
-    ) -> list[dict[str, Any]]:
-        """List all rows for a specific configuration.
-
-        Args:
-            component_id: Component identifier (e.g. 'keboola.ex-http').
-            config_id: Configuration ID.
-            branch_id: If set, target a specific dev branch.
-
-        Returns:
-            List of config row dicts.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        resp = self._request(
-            "GET",
-            f"{prefix}/components/{quote(component_id)}/configs/{quote(config_id)}/rows",
-        )
-        return resp.json()
-
-    def get_config_row(
-        self,
-        component_id: str,
-        config_id: str,
-        row_id: str,
-        branch_id: int | None = None,
-    ) -> dict[str, Any]:
-        """Get a single configuration row by ID.
-
-        Args:
-            component_id: Component identifier.
-            config_id: Configuration ID.
-            row_id: Row ID.
-            branch_id: If set, target a specific dev branch.
-
-        Returns:
-            Row detail dict from the API.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        resp = self._request(
-            "GET",
-            f"{prefix}/components/{quote(component_id)}/configs/{quote(config_id)}/rows/{quote(row_id)}",
-        )
-        return resp.json()
-
-    def get_config_detail(
-        self,
-        component_id: str,
-        config_id: str,
-        branch_id: int | None = None,
-    ) -> dict[str, Any]:
-        """Get detailed information about a specific configuration.
-
-        Args:
-            component_id: The component ID (e.g. keboola.ex-db-snowflake).
-            config_id: The configuration ID.
-            branch_id: If set, get detail from a specific dev branch.
-
-        Returns:
-            Configuration detail dict from the API.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        safe_component_id = quote(component_id, safe="")
-        safe_config_id = quote(config_id, safe="")
-        response = self._request(
-            "GET",
-            f"{prefix}/components/{safe_component_id}/configs/{safe_config_id}",
-        )
-        return response.json()
-
-    def get_config_state(
-        self,
-        component_id: str,
-        config_id: str,
-        branch_id: int | None = None,
-    ) -> dict[str, Any]:
-        """Get the runtime state dict of a specific configuration.
-
-        Convenience wrapper over
-        ``get_config_detail(...).get("state", {})``: Storage API does not
-        expose a standalone ``GET .../state`` resource (production returns
-        404, branch-scoped returns 501 Not Implemented), so the state is
-        only served inline as a field inside the configuration detail
-        response. This wrapper is retained for API discoverability, but
-        callers that already have a detail response should read ``state``
-        from it directly instead of issuing this second identical request
-        -- the service layer's single-mode ``--with-state`` does exactly
-        that (see ``ConfigService.get_config_detail``).
-
-        For bulk state retrieval across many configs, prefer the
-        ``include=state`` query param on
-        ``list_components_with_configs(include="configuration,rows,state")``
-        -- one request serves every config's state instead of N requests.
-
-        Args:
-            component_id: The component ID (e.g. keboola.ex-db-snowflake).
-            config_id: The configuration ID.
-            branch_id: If set, fetch state from a specific dev branch.
-
-        Returns:
-            The state dict (empty ``{}`` when the config has no saved state).
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        safe_component_id = quote(component_id, safe="")
-        safe_config_id = quote(config_id, safe="")
-        response = self._request(
-            "GET",
-            f"{prefix}/components/{safe_component_id}/configs/{safe_config_id}",
-        )
-        body = response.json()
-        state = body.get("state")
-        return state if isinstance(state, dict) else {}
-
-    def list_config_folder_metadata(self, branch_id: int) -> dict[str, str]:
-        """Fetch folder names for all configurations via metadata search.
-
-        Uses the search/component-configurations endpoint to find configs
-        with ``KBC.configuration.folderName`` metadata.
-
-        Note: This endpoint requires a branch ID (branch-only route).
-
-        Args:
-            branch_id: Branch ID (required — use default branch for production).
-
-        Returns:
-            Dict mapping ``"{component_id}/{config_id}"`` to folder name.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}"
-        resp = self._request(
-            "GET",
-            f"{prefix}/search/component-configurations",
-            params={
-                "metadataKeys[]": "KBC.configuration.folderName",
-                "include": "filteredMetadata",
-            },
-        )
-        folder_map: dict[str, str] = {}
-        for item in resp.json():
-            comp_id = item.get("idComponent", "")
-            config_id = str(item.get("configurationId", ""))
-            meta = next(
-                (m for m in item.get("metadata", []) if m["key"] == "KBC.configuration.folderName"),
-                None,
-            )
-            if meta:
-                folder_map[f"{comp_id}/{config_id}"] = meta["value"]
-        return folder_map
-
-    def list_config_metadata(
-        self,
-        component_id: str,
-        config_id: str,
-        branch_id: int | None = None,
-    ) -> list[dict[str, Any]]:
-        """List metadata entries on a configuration.
-
-        GET /v2/storage/[branch/{b}/]components/{c}/configs/{id}/metadata
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        response = self._request(
-            "GET",
-            f"{prefix}/components/{quote(component_id, safe='')}/configs/{quote(config_id, safe='')}/metadata",
-        )
-        return response.json()
-
-    def set_config_metadata(
-        self,
-        component_id: str,
-        config_id: str,
-        entries: list[tuple[str, str]],
-        branch_id: int | None = None,
-    ) -> list[dict[str, Any]]:
-        """Bulk-set metadata key/value pairs on a configuration.
-
-        POST /v2/storage/[branch/{b}/]components/{c}/configs/{id}/metadata
-        Same PHP-style indexed form as set_branch_metadata.
-        """
-        form: dict[str, str] = {}
-        for i, (key, value) in enumerate(entries):
-            form[f"metadata[{i}][key]"] = key
-            form[f"metadata[{i}][value]"] = value
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        response = self._request(
-            "POST",
-            f"{prefix}/components/{quote(component_id, safe='')}/configs/{quote(config_id, safe='')}/metadata",
-            data=form,
-        )
-        return response.json()
-
-    def delete_config_metadata(
-        self,
-        component_id: str,
-        config_id: str,
-        metadata_id: int | str,
-        branch_id: int | None = None,
-    ) -> None:
-        """Delete a single metadata entry on a configuration by its numeric ID.
-
-        DELETE /v2/storage/[branch/{b}/]components/{c}/configs/{id}/metadata/{mid}
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        self._request(
-            "DELETE",
-            f"{prefix}/components/{quote(component_id, safe='')}/configs/{quote(config_id, safe='')}/metadata/{metadata_id}",
-        )
-
-    def create_config(
-        self,
-        component_id: str,
-        name: str,
-        configuration: dict[str, Any],
-        description: str = "",
-        branch_id: int | None = None,
-        is_disabled: bool = False,
-    ) -> dict[str, Any]:
-        """Create a new configuration for a component.
-
-        POST /v2/storage/[branch/{id}/]components/{comp_id}/configs
-
-        Args:
-            component_id: Component identifier.
-            name: Configuration name.
-            configuration: Configuration body (parameters, storage, etc.).
-            description: Optional description.
-            branch_id: If set, target a specific dev branch.
-            is_disabled: When True, the configuration is created in disabled
-                state (mirrors ``create_config_row``).
-
-        Returns:
-            Created configuration dict including the assigned 'id'.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        data: dict[str, Any] = {
-            "name": name,
-            "description": description,
-            "configuration": json.dumps(configuration),
-        }
-        if is_disabled:
-            data["isDisabled"] = "1"
-        resp = self._request(
-            "POST",
-            f"{prefix}/components/{quote(component_id)}/configs",
-            data=data,
-        )
-        return resp.json()
-
-    def update_config(
-        self,
-        component_id: str,
-        config_id: str,
-        name: str | None = None,
-        configuration: dict[str, Any] | None = None,
-        description: str | None = None,
-        change_description: str = "",
-        branch_id: int | None = None,
-        is_disabled: bool | None = None,
-    ) -> dict[str, Any]:
-        """Update an existing configuration.
-
-        PUT /v2/storage/[branch/{id}/]components/{comp_id}/configs/{config_id}
-
-        Only provided (non-None) fields are sent in the request.
-        ``is_disabled=None`` leaves the remote enabled/disabled state
-        untouched (mirrors ``update_config_row``).
-
-        Returns:
-            Updated configuration dict.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        data: dict[str, Any] = {}
-        if name is not None:
-            data["name"] = name
-        if description is not None:
-            data["description"] = description
-        if configuration is not None:
-            data["configuration"] = json.dumps(configuration)
-        if is_disabled is not None:
-            data["isDisabled"] = "1" if is_disabled else "0"
-        if change_description:
-            data["changeDescription"] = change_description
-        resp = self._request(
-            "PUT",
-            f"{prefix}/components/{quote(component_id)}/configs/{quote(config_id)}",
-            data=data,
-        )
-        return resp.json()
-
-    def create_config_row(
-        self,
-        component_id: str,
-        config_id: str,
-        name: str,
-        configuration: dict[str, Any],
-        description: str = "",
-        is_disabled: bool = False,
-        branch_id: int | None = None,
-    ) -> dict[str, Any]:
-        """Create a new configuration row.
-
-        POST /v2/storage/[branch/{id}/]components/{comp_id}/configs/{config_id}/rows
-
-        Args:
-            component_id: The component ID.
-            config_id: The parent configuration ID.
-            name: Row name.
-            configuration: Row-level configuration dict.
-            description: Optional row description.
-            is_disabled: When True, the row is created in disabled state and
-                excluded from job runs until re-enabled.
-            branch_id: Optional dev branch ID.
-
-        Returns:
-            Created row dict including the assigned 'id'.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        data: dict[str, Any] = {
-            "name": name,
-            "description": description,
-            "configuration": json.dumps(configuration),
-        }
-        if is_disabled:
-            data["isDisabled"] = "1"
-        resp = self._request(
-            "POST",
-            f"{prefix}/components/{quote(component_id)}/configs/{quote(config_id)}/rows",
-            data=data,
-        )
-        return resp.json()
-
-    def update_config_row(
-        self,
-        component_id: str,
-        config_id: str,
-        row_id: str,
-        name: str | None = None,
-        configuration: dict[str, Any] | None = None,
-        description: str | None = None,
-        is_disabled: bool | None = None,
-        change_description: str = "",
-        branch_id: int | None = None,
-    ) -> dict[str, Any]:
-        """Update an existing configuration row.
-
-        PUT /v2/storage/[branch/{id}/]components/{comp_id}/configs/{config_id}/rows/{row_id}
-
-        Args:
-            is_disabled: When True, disable the row; when False, enable it;
-                when None, leave the current state unchanged.
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        data: dict[str, Any] = {}
-        if name is not None:
-            data["name"] = name
-        if description is not None:
-            data["description"] = description
-        if configuration is not None:
-            data["configuration"] = json.dumps(configuration)
-        if is_disabled is not None:
-            data["isDisabled"] = "1" if is_disabled else "0"
-        if change_description:
-            data["changeDescription"] = change_description
-        resp = self._request(
-            "PUT",
-            f"{prefix}/components/{quote(component_id)}/configs/{quote(config_id)}/rows/{quote(row_id)}",
-            data=data,
-        )
-        return resp.json()
-
-    def delete_config_row(
-        self,
-        component_id: str,
-        config_id: str,
-        row_id: str,
-        branch_id: int | None = None,
-    ) -> None:
-        """Delete a configuration row.
-
-        DELETE /v2/storage/[branch/{id}/]components/{comp_id}/configs/{config_id}/rows/{row_id}
-        """
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        self._request(
-            "DELETE",
-            f"{prefix}/components/{quote(component_id)}/configs/{quote(config_id)}/rows/{quote(row_id)}",
-        )
 
     def list_buckets(
         self, include: str | None = None, branch_id: int | None = None
@@ -1986,273 +1506,3 @@ class KeboolaClient(
                 shutil.copyfileobj(source, fh, length=FILE_DOWNLOAD_CHUNK_SIZE)
 
         return out_path.stat().st_size
-
-    def list_jobs(
-        self,
-        component_id: str | None = None,
-        config_id: str | None = None,
-        status: str | None = None,
-        limit: int = DEFAULT_JOB_LIMIT,
-        offset: int = 0,
-    ) -> list[dict[str, Any]]:
-        """List jobs from the Queue API.
-
-        Args:
-            component_id: Optional filter by component ID.
-            config_id: Optional filter by config ID (requires component_id).
-            status: Optional filter by job status.
-            limit: Max number of jobs to return (1-500).
-            offset: Offset for pagination.
-
-        Returns:
-            List of job dicts from the Queue API.
-        """
-        params: dict[str, str | int] = {"limit": limit, "offset": offset}
-        if component_id:
-            params["component"] = component_id
-        if config_id:
-            params["config"] = config_id
-        if status:
-            params["status"] = status
-
-        response = self._queue_request("GET", "/search/jobs", params=params)
-        return response.json()
-
-    def list_jobs_grouped(
-        self,
-        jobs_per_group: int = DEFAULT_JOBS_PER_CONFIG,
-        limit: int = DEFAULT_GROUPED_JOBS_LIMIT,
-        sort_by: str = "startTime",
-        sort_order: str = "desc",
-        created_time_from: str | None = None,
-    ) -> list[dict[str, Any]]:
-        """List jobs grouped by component+config from the Queue API.
-
-        Uses GET /search/grouped-jobs to fetch the latest N jobs for each
-        unique component+config combination in a single API call.
-
-        Args:
-            jobs_per_group: Max jobs per component+config group (1-500).
-            limit: Max number of groups to return (1-500).
-            sort_by: Sort field for jobs within each group.
-            sort_order: Sort direction ("asc" or "desc").
-            created_time_from: Optional ISO datetime filter (e.g. "2026-03-20T00:00:00Z").
-
-        Returns:
-            List of group dicts: [{"group": {"componentId": ..., "configId": ...}, "jobs": [...]}]
-        """
-        params: list[tuple[str, str]] = [
-            ("groupBy[]", "componentId"),
-            ("groupBy[]", "configId"),
-            ("jobsPerGroup", str(jobs_per_group)),
-            ("limit", str(limit)),
-            ("sortBy", sort_by),
-            ("sortOrder", sort_order),
-        ]
-        if created_time_from:
-            params.append(("filters[createdTimeFrom]", created_time_from))
-
-        response = self._queue_request("GET", "/search/grouped-jobs", params=params)
-        return response.json()
-
-    def get_job_detail(self, job_id: str) -> dict[str, Any]:
-        """Get detailed information about a specific job from the Queue API.
-
-        Args:
-            job_id: The job ID.
-
-        Returns:
-            Job detail dict from the Queue API.
-        """
-        safe_job_id = quote(job_id, safe="")
-        response = self._queue_request("GET", f"/jobs/{safe_job_id}")
-        return response.json()
-
-    # --- Queue Job Creation ---
-
-    def create_job(
-        self,
-        component_id: str,
-        config_id: str,
-        config_data: dict[str, Any] | None = None,
-        config_row_ids: list[str] | None = None,
-        mode: str = "run",
-        branch_id: int | None = None,
-        variable_values_id: str | None = None,
-    ) -> dict[str, Any]:
-        """Create and run a Queue API job.
-
-        Args:
-            component_id: Component ID (e.g. keboola.sandboxes).
-            config_id: Configuration ID.
-            config_data: Optional runtime config data override.
-            config_row_ids: Optional list of config row IDs to run
-                (omit to run entire config).
-            mode: Job mode (default: run).
-            branch_id: Optional dev branch ID. When set, the job runs
-                on that branch instead of the default (production) branch.
-            variable_values_id: Optional id of a row in the linked
-                ``keboola.variables`` config. When set, the Queue API binds
-                the row's values to the job's `{{ variable }}` placeholders.
-                Omit for configurations that have no linked variables.
-
-        Returns:
-            Job dict from the Queue API.
-        """
-        body: dict[str, Any] = {
-            "component": component_id,
-            "config": config_id,
-            "mode": mode,
-        }
-        if branch_id is not None:
-            body["branchId"] = str(branch_id)
-        if config_data:
-            body["configData"] = config_data
-        if config_row_ids:
-            body["configRowIds"] = config_row_ids
-        if variable_values_id:
-            body["variableValuesId"] = variable_values_id
-        response = self._queue_request("POST", "/jobs", json=body)
-        return response.json()
-
-    def kill_job(self, job_id: str) -> dict[str, Any]:
-        """Request termination of a running Queue API job.
-
-        Sets the job's desiredStatus to "terminating"; the executor transitions
-        the actual status asynchronously (waiting -> cancelled, processing ->
-        terminating -> terminated). Poll get_job_detail until isFinished=True
-        to observe the terminal state.
-
-        Killable states per Queue API: created, waiting, processing. Calling
-        kill on any other state returns HTTP 400 with a "not in one of killable
-        states" message; callers that want idempotent behavior (e.g. bulk
-        terminate after list_jobs under race conditions) should translate that
-        into a no-op success at the service layer.
-        """
-        safe_job_id = quote(job_id, safe="")
-        response = self._queue_request("POST", f"/jobs/{safe_job_id}/kill")
-        return response.json()
-
-    def fetch_job_events(self, run_id: str, limit: int | None = None) -> list[dict[str, Any]]:
-        """Fetch events emitted during a job's run.
-
-        Wraps the Storage API's ``GET /v2/storage/events?runId={runId}``
-        endpoint -- NOT a Queue API path. Queue jobs (Queue API v2) expose a
-        ``runId`` on the job dict (typically equal to the job ``id``); the
-        Storage Events API is the canonical event feed for the job. Returns
-        the list in Storage API order (newest -> oldest; callers that want
-        a chronological "tail" should reverse the slice).
-
-        Args:
-            run_id: The job's ``runId`` (``job["runId"]``; falls back to
-                ``job["id"]`` on legacy records where they match).
-            limit: Optional server-side event cap. Storage API default is
-                about 100; pass an explicit value to cover long runs.
-
-        Returns:
-            List of event dicts. Each event typically has ``uuid``,
-            ``event``, ``component``, ``message``, ``type``, ``created``,
-            ``runId``, ``configurationId`` keys. Empty when the run emitted
-            no events yet.
-        """
-        params: dict[str, Any] = {"runId": run_id}
-        if limit is not None and limit > 0:
-            params["limit"] = limit
-        response = self._request("GET", "/v2/storage/events", params=params)
-        payload = response.json()
-        # Storage events returns a bare list. Tolerate a dict-wrapped
-        # future shape defensively.
-        if isinstance(payload, list):
-            return payload
-        if isinstance(payload, dict) and isinstance(payload.get("events"), list):
-            return payload["events"]
-        return []
-
-    def wait_for_queue_job(
-        self,
-        job_id: str,
-        max_wait: float = STORAGE_JOB_MAX_WAIT,
-        poll_strategy: str = DEFAULT_POLL_STRATEGY,
-    ) -> dict[str, Any]:
-        """Poll a Queue API job until it reaches a terminal state.
-
-        Uses the piecewise ``JOB_POLL_CURVE`` from constants for the
-        ``"exponential"`` strategy (2s x 30 -> 5s x 48 -> 15s forever) and
-        the legacy fixed ``STORAGE_JOB_POLL_INTERVAL`` for ``"fixed"``. The
-        curve matches the cadence used by FIIA and the official
-        ``keboola-as-code`` Go CLI.
-
-        Args:
-            job_id: The Queue job ID.
-            max_wait: Maximum seconds to wait (default: STORAGE_JOB_MAX_WAIT).
-            poll_strategy: "exponential" (default) or "fixed". Any other
-                value raises ValueError before the first network call.
-
-        Returns:
-            Completed job dict.
-
-        Raises:
-            ValueError: If poll_strategy is not one of VALID_POLL_STRATEGIES.
-            KeboolaApiError: If the job fails (QUEUE_JOB_FAILED) or the
-                deadline elapses before the job finishes (QUEUE_JOB_TIMEOUT).
-        """
-        if poll_strategy not in VALID_POLL_STRATEGIES:
-            # ValueError (not KeboolaApiError) because this is a programming
-            # error -- the caller passed an invalid literal, not a bad API
-            # response. JobService validates before reaching this layer, so
-            # hitting this path from the CLI would be a bug in kbagent.
-            raise ValueError(
-                f"Invalid poll_strategy {poll_strategy!r}. "
-                f"Expected one of: {sorted(VALID_POLL_STRATEGIES)}."
-            )
-
-        deadline = time.monotonic() + max_wait
-        for interval in _iter_poll_intervals(poll_strategy):
-            job = self.get_job_detail(job_id)
-            if job.get("isFinished"):
-                if job.get("status") == "error":
-                    result = job.get("result", {})
-                    error_msg = (
-                        result.get("message", "Queue job failed")
-                        if isinstance(result, dict)
-                        else "Queue job failed"
-                    )
-                    raise KeboolaApiError(
-                        message=f"Queue job {job_id} failed: {error_msg}",
-                        status_code=500,
-                        error_code=ErrorCode.QUEUE_JOB_FAILED,
-                        retryable=False,
-                    )
-                return job
-
-            # Cap the sleep so we never blow past the deadline by more than
-            # one interval: trim to whatever time remains; if zero, break.
-            remaining = deadline - time.monotonic()
-            if remaining <= 0:
-                break
-            time.sleep(min(interval, remaining))
-
-        raise KeboolaApiError(
-            message=f"Queue job {job_id} did not complete within {max_wait}s",
-            status_code=504,
-            error_code=ErrorCode.QUEUE_JOB_TIMEOUT,
-            retryable=True,
-        )
-
-    def delete_config(
-        self, component_id: str, config_id: str, branch_id: int | None = None
-    ) -> None:
-        """Delete a component configuration.
-
-        Args:
-            component_id: Component ID.
-            config_id: Configuration ID.
-            branch_id: Branch ID. If provided, deletes config in that branch.
-        """
-        safe_component = quote(component_id, safe="")
-        safe_config = quote(config_id, safe="")
-        prefix = f"/v2/storage/branch/{branch_id}" if branch_id else "/v2/storage"
-        self._request(
-            "DELETE",
-            f"{prefix}/components/{safe_component}/configs/{safe_config}",
-        )
